NAS-136152 / 25.10 / Move private CRUD service models out of the versioned API (#16582)
We decided a while ago not to version private methods, but a few models
slipped into the public API anyway. The ones addressed in this PR
include:

* `rdma.interface`
* `acme.registration`
* `fcport` incorrectly marked as private (used in webui)
* `fc.fc_host` incorrectly marked as private (used in webui)

By passing `_config.private` to `api_method` in
`CRUDServiceMetabase.__new__`, we can be a little more careful about
making this mistake in the future. We should examine a similar change
for `ConfigService`.