NAS-135200 / 25.10 / Add support for directory services authentication for FTP. (#16598)
Active Directory users are not able to authenticate via the FTP
protocol.

This PR fixes that with a configuration change in the `proftpd.conf`
file when a directory service is enabled.

This has been manually tested against an AD server. 

Testing against a FreeIPA server and CI tests will be implemented in a
separate PR.