NAS-136155 / 25.10 / Convert sharing.smb to pydantic schema (#16584)
This commit refactors the sharing.smb CRUD endpoint to use a new
pydantic schema with a discriminator on the share `purpose` key. Options
are linked explicitly with the share's purpose via a new `options`
field. Fields for SMB share are now based on a strenum value wherever
possible to improve ability for linter to pick up errors. A limited
scope compatibilty wrapper is left in place to avoid totally breaking
the SMB share UI until webui team can accommodate changes.