Fix configuration for SMB Veeam Fast Clone
This commit adds a new SMB share purpose of VEEAM_REPOSITORY_SHARE,
which requires the ZFS recordsize to be set to 128 KiB (per guidance
from Veeam) and sets the underlying SMB block size to 128 KiB so
that FileFsSizeInformation the Sectors Per Allocation Unit is
reported to Veeam Backup & Restore in such a way that it will use
the correct offsets and lengths when it issues
FSCTL_DUPLICTE_EXTENTS_TO_FILE requests. If this is not done, then
Fast Copy requests will be rejected by ZFS due to improper alignment
and ZFS will fallback to doing an internal copy of the specified
range.