Handle unlicensed users edge case for installing apps
This commit adds changes for a case where a user with a license was able to install apps but then had his license updated and then he was not allowed but was still able to install apps. This makes sure that in such cases if now a user is unlicensed, he is not able to view/operate any of the existing apps.