Do not crash in pool.dataset.processes on procfd EACCES
This addresses an issue reported by community member in the forums.
os.stat("/proc/<pid>/<fd>") failed with PermissionError in case where
pid was for nextcloud snap installed in Ubuntu container and readlink
on fd was `/snap/nextcloud/48223/bin/run-httpd`. Suppressing the
error in this situation is okay since the snap will always be
installed in the root dataset of the container and so check for
running containers is sufficient for our purposes of identifying
in-use datasets.

(cherry picked from commit 0e3e329a33f37e4f1517bd8f1cef33d248a31d91)