NAS-136328 / 25.04.2 / Remove TOKEN_PLAIN from auth mech choices for OTPW (by anodos325) (#16644)
This is an enhancement requested by the UI team. If the current session
is authenticated via a onetime password, adjust our available
authentication mechanism choices for the *session* to exclude
authentication tokens. This provides a hint to UI to not issue request
to generate the token and allows code simplification.

Original PR: https://github.com/truenas/middleware/pull/16636

Co-authored-by: Andrew Walker <awalker@ixsystems.com>