NAS-136428 / 25.10 / fix QuotaAlertSource (#16652)
Recent optimizations here introduced a subtle regression found by QE
team. The problem is that zfs user properties are always strings but for
our use-case, we need to cast some of then to integers so we can have
this alert function as desired. This fixes the issue by casting the
specific zfs user properties we're after.