NAS-135915 / 25.10 / Don't treat NSS UNAVAIL return code as fatal when building cache (#16651)
This commit changes how the UNAVAIL return code is treated while
building cache for SSSD when joined to IPA / LDAP. It may take some time
for SSSD to initialize and build its own caches. In this case we can
squash the NSS error and sleep a bit more to ensure we don't have
failures.