Remove ID type both and version directory services cache
This commit removes the `id_type_both` key from user entries since we
are not abstracting away this implementation detail from from the
public API in SMB share ACL and other areas. This commit also adds
a version identifier to directory services caches so that we
automatically invalidate them on middleware startup of the underlying
TrueNAS version changed (this is because the fields for users can
change between versions).