NAS-136439 / 25.10 / Remove ID type both and version directory services cache (#16659)
This commit removes the `id_type_both` key from user entries since we
are not abstracting away this implementation detail from from the public
API in SMB share ACL and other areas. A side-effect of this change is
that idmap domain information no longer needs to be queried when
generating AD cache (or passed to related functions). This commit also
adds a version identifier to directory services caches so that we
automatically invalidate them on middleware startup of the underlying
TrueNAS version changed (this is because the fields for users can change
between versions).