Handle KRBError when trying to restore secrets
This commit fixes some verbose log spam that was being generated
when middleware tried to recover from someone deleting our
computer account from under us. We failed to kinit via keytab and
our stored machine account password in the local secrets.tdb file,
and so we tried to restore a saved copy of it. During the restoration
we attempt to kinit with the old credential, and this was failing
due with a KRBError that wasn't being caught and turned into an
ADHealthError. This commit adds some minimal decoding of the
kerberos error in the context of our stored machine account and
eliminates the log spam.