NAS-135866 / 25.10 / Remove unwanted newlines in API docs (#16525)
The following field docstring:
```python
    field: int
    """Just a
    chill guy."""
```
produces the field description `"Just a\nchill guy."` which isn't
pleasant to look at in the API docs.

## Rejected Solution
Parse double newlines as single newlines and single newlines like in the
example above as spaces.

Using this implicit docstring parsing makes some formatting unintuitive
or unwieldy, particularly when using reST in API docstrings.

## Solution
Instead, we need to be intentional when writing API docstrings by ending
line wraps with backslashes to remove the newline. So the above example
would become:
```python
    field: int
    """Just a \
    chill guy."""
```
and would parse to `"Just a chill guy."` as intended.

## Proposed Standards

Now that our API docs are [provided to the
public](https://api.truenas.com/v25.10.0/), we need to establish some
formatting guidelines. The changes made in this PR include:
* Only use triple quotation marks `"""` for API docstrings as per
[PEP-257](https://peps.python.org/pep-0257/#:~:text=For%20consistency%2C%20always%20use%20%22%22%22triple%20double%20quotes%22%22%22).
Triple apostrophes `'''`, single quotation marks `"`, and single
apostrophes `'` should not be used to denote docstrings.
* API docstrings should begin with a capital letter unless the first
word is a variable name or other identifier.
* API docstrings all must end with a period.