NAS-135988 / 25.10 / simplify pool.dataset.details (#16669)
With the optimization in
https://github.com/truenas/middleware/pull/16628, we're able to simplify
this endpoint. This also fixes some bugs where-by the atime property
wasn't being properly handled when it was inherited.

Before my changes, these properties were formatted like so:
```
    "atime": false,
    "casesensitive": true,
    "readonly": false,
```

With my changes they are formatted like so (which matches the format of
all the other zfs properties as well)
```
    "atime": {
      "parsed": false,
      "rawvalue": "off",
      "value": "OFF",
      "source": "LOCAL",
      "source_info": null
    },
    "readonly": {
      "parsed": false,
      "rawvalue": "off",
      "value": "OFF",
      "source": "DEFAULT",
      "source_info": null
    },
    "casesensitivity": {
      "parsed": "sensitive",
      "rawvalue": "sensitive",
      "value": "SENSITIVE",
      "source": "NONE",
      "source_info": null
    }
```