NAS-136426 / 25.10 / Allow specifying interfaces for TrueNAS connect (#16686)
This PR adds changes to allow binding interfaces with truenas connect.
Basically this will mean that users don't have to rely on specifying IPs
which can change later on and rather we will use interfaces which are
persistent and remain.

To catch up to any ip changes on interfaces, we will be doing that in a
subsequent PR.