NAS-136596 / 25.10 / Add automatic interface discovery for TrueNAS Connect (#16712)
This PR adds changes so by default TNC is configured with all the
available interfaces on the system. What this means in practice is that
whatever IPs we have configured on any interface, we will send them to
TNC so their A records can be created appropriately unless user
explicitly disables this with his/her own personal preference.