Adjust job progress when applying update file
This commit fixes an issue where if scale build gives 100 or 96 percentage, we walk back from that figure when in middleware we send a subsequent event with 95 percentage. So in such cases, we max out the progress reported by scale build to 90 and then take it to completion in middleware.