NAS-136660 / 25.10 / Properly validate ports in virt plugin (#16721)
## Problem
The `port_details` field in `port.ports_mapping` includes descriptions
for all ports used by a service. Currently, the code retrieves and
displays the descriptions of all ports without filtering for the
specific port being used by the application.

## Solution

Update the logic to filter and include only the description relevant to
the specific port used by the application.