NAS-136578 / 25.10 / Disable restart limiter for syslog-ng (#16727)
The syslog-ng tests were activating the systemd restart limiter.
Disabling syslog-ng resulted in a cascade of failures for the audit CI
tests.

To prevent the CI testing failures and avoid the possibility of systemd
disabling syslog-ng in runtime, this PR disables the restart limiter for
syslogn-ng.


This was tested with 6 CI runs, three with single and 3 with HA. None
exhibited the audit failure cascade.