Ensure ix-apps is mounted at it's proper place
This commit fixes a bug where if a user rollsback to a TN version which does not has docker apps, ix-apps dataset mountpoint would be reset and it would get mounted under it's root dataset. In this case we would like to make sure that the mountpoint correctly reflects where we want ix-apps to be mounted.