NAS-136094 / 25.10 / Avoid unnecessary cache refresh on failover events (#16747)
This commit shifts the directory services cache files to the system
dataset and adds an expiration key to the caches so that middleware can
be more intelligent about whether to attempt to rebuild the cache.

This happens in one of the following circumstances:
1. User initiates cache rebuild
2. User just joined a directory service
3. Cronjob issues cache refresh per crontab

The ultimate effect is that `directoryservices.setup` during a failover
event to become active will no longer trigger an expensive cache rebuild
and put unnecessary load on winbindd or sssd.