Persist KDC affinity across HA failovers
Replication between KDCs can be slow (especially in our virtualized
domains), and so we set a one hour affinity for a specific KDC
after performing certain administrative operations. Originally the
affinity information was written to the middleware run directory,
but this is not ideal for HA because it's conceivable that the
administrator will immediately test failovers after joining AD.

This commit changes the timeout so that its based on real time rather
than monotonic time and shifts the file to directory_services directory
in the system dataset.