NAS-136756 / 25.10 / Add ability to specify primary_kdc for realms (#16768)
This commit allows the administrator to specify a primary_kdc for
kerberos realms. The primary kdc is used by the kerberos client when
acquiring creds if the KDC it's currently using fails to with bad
password. This is valuable for domains with a hub-and-spoke topology in
which password changes may slowly propogate from the hub to the spoke.