NAS-136773 / 25.10 / Fix broken upgrades due to ctypes failure (#16778)
During update.run we chroot into the squashfs filesystem to perform
various operations. This causes attempts to generate python-pam handles
to fail. Rather than trying to massage ctypes to find the libraries
within the squashfs filesystem, we can just set the PAM_REF to None. The
installer shouldn't create middleware pam handles and so this is a
low-risk solution to the problem.