NAS-136775 / 25.04.2 / fallthrough auth.login_ex on unhealthy HA (by yocalebo) (#16784)
Seems to be a reoccurring theme whereby HA is disabled and then both
controllers are power cycled, rebooted, etc. When HA is
disabled...failover will not work by design. It has been this way since
the beginning. However, some changes to the UI login page have
introduced a situation where the user is unable to login at all without
any indication as to why (even if the credentials that are provided are
valid).

This changes the logic to return REDIRECT _ONLY_ when HA is healthy and
the user is trying to login to the standby controller. Otherwise, the
login attempt will fall through like normal.

This shifts the responsibility of whether to continue the login process
and show the web page contents back to the caller (local UI in this
scenario).

Original PR: https://github.com/truenas/middleware/pull/16781

Co-authored-by: Caleb <yocalebo@gmail.com>