NAS-136771 / 25.10 / Factor in partitions as well when generating disk identifier (#16777)
## Problem

Currently how disk identifier is calculated in `disk.query`, it factors
in partition uuids as well if they are available and serial/lunid are
not before falling back to using disk name as disk identifier. This
introduces discrepancy between identifiers wherever they are going to be
used elsewhere.

## Solution

Factor in partition uuid when determining disk identifier in `DiskEntry`
class to make sure that both implementations are as close as possible in
terms of retrieving the same deterministic disk identifier.