NAS-136666 / 25.10 / Account for partition start in `disk.get_data_partition_size` (#16793)
There are some systems with pools created on legacy systems. That
systems have data partitions starting after swap partition. Account for
that.

`test_expand_partition_does_not_destroy_data` was testing a situation
that is no longer possible (we won't be trying to expand beyond the disk
size if the expanded partition does not start at the beginning of the
disk).