NAS-136653 / 25.10 / Update GPU PCI choices to reflect which GPU is critical (#16787)
## Problem

Currently `system.advanced.get_gpu_pci_choices` is showing GPU slots
which are not critical and can be isolated from the system. However this
means that the user in the UI sometimes is not aware why a GPU which the
system has is not showing up and why he cannot isolate it.

## Solution

Update the endpoint UI uses so we report the GPU with the PCI slot which
UI should use so middleware can isolate a GPU but also send another flag
which shows if the GPU is critical or not. User should not be able to
select this as an option but he can at least see that it is critical.

i.e
```
❯ midclt call system.advanced.get_gpu_pci_choices | jq
{
  "Red Hat, Inc. QXL paravirtual graphic card [0000:00:02.0]": {
    "pci_slot": "0000:00:02.0",
    "uses_system_critical_devices": true,
    "critical_reason": "Critical devices found: 0000:00:01.0, 0000:00:00.0"
  }
}
```