NAS-136831 / 25.10 / fix virtio-blk serial number detection (#16799)
We're no longer properly detecting serial numbers for virtio-blk based
disks. This is because the `serial` file is put at the root of the block
device directory in sysfs. Changes here just add another path check to
ensure we're grabbing the serial number properly. This was confirmed to
work on another developer's VM.