NAS-136864 / 25.10 / fix 2fa on HA (#16810)
After 2fa is enabled on HA, a failover would lead to a situation where
the login page would not present the 2fa form. This is because we're
only generating the oath file on the currently running active. Fix the
issue by going through the `service.control` API which ensures the
operations will automatically propagate to the standby on HA systems.