Regenerate the users.oath file when DS becomes healthy
Active directory user oath configuration is keyed to the user SID
value rather than being based on name, uid, or synthetic datastore
IDs because all of the former are mutable. This means that if
AD is broken when the users.oath file is generated then AD users
will be omitted. When we transition from FAULTED to HEALTHY through
directoryservices.health.recover we need to regenerate the users.oath
file to properly resolve any directoryservices entries.