NAS-136920 / 25.04.2.1 / fix busy loop on HA wrt to service.* actions (by yocalebo) (#16837)
On an internal system, an HA system was exposed to a network
segmentation between the controllers. This caused both controllers to
become MASTER. This caused a busy-loop of core.bulk service.restart
calls to be sent back and forth to each controller endlessly. To remedy
this situation I've added an extra check to ensure that
`failover.disabled.reasons` needs to be empty before forwarding the
remote call.

Original PR: https://github.com/truenas/middleware/pull/16836

---------

Co-authored-by: Caleb <yocalebo@gmail.com>