NAS-136949 / 25.10 / Ignore libvirt service actions in service_remote on HA systems (#16838)
This is the "belt and suspenders" approach that compliments the changes
introduced here https://github.com/truenas/middleware/pull/16830. With
those changes this shouldn't be possible, however, it's better to be
safe than sorry. While I'm here, I went ahead and added some comments to
explain what each if-condition does and it's context.