NAS-136941 / 25.10 / Fix display device password/port validation (#16839)
## Problem

We have users who had old VMs where `web_port` could be 5800 and also
`passoword` as an optional field. Around bluefin, we changed this
behaviour to make sure that ports below 5900 cannot be used for new
display devices and that display devices require a password.
However with the change to pydantic, these validations got moved to
query as well which breaks these existing consumers.

## Solution

Make sure we do not validate this data when we are querying these
devices as we don't want to break existing consumers.