NAS-136989 / 25.10-RC.1 / Fix cdrom device path validation (#16859)
## Problem
In older releases we allowed CDROM devices to live outside `/mnt` which
is not ideal. However the problem we see right now is that if we
validate this in the pydantic model itself, `vm.query` can fail which is
not ideal.

## Solution
We are already checking that the cdrom device is in a pool, so we don't
have to do the same validation in pydantic.