NAS-136989 / 26.04 / Fix cdrom device path validation (by Qubad786) (#16863)
## Problem
In older releases we allowed CDROM devices to live outside `/mnt` which
is not ideal. However the problem we see right now is that if we
validate this in the pydantic model itself, `vm.query` can fail which is
not ideal.

## Solution
We are already checking that the cdrom device is in a pool, so we don't
have to do the same validation in pydantic.

Original PR: https://github.com/truenas/middleware/pull/16859

---------

Co-authored-by: M. Rehan <mrehanlm93@gmail.com>