NAS-136994 / 25.04.2.1 / Do not configure directoryservices nsswitch on standby (#16864)
This commit ensures that directoryservices do not get inserted into the
nsswitch.conf on standby controller if directoryservices are enabled.
25.04 and earlier expect the standby controller to not be fully
configured for directory services and so we don't want NSS calls to
potentially block on waiting for winbindd / sssd to bind to
non-configured service. When vrrp_master gets called the nsswitch.conf
is automatically reconfigured during a middleware call to
directoryservices.setup.