NAS-136902 / 25.10-BETA.1 / change update profile verbiage (by yocalebo) (#16887)
We were asked to change the new update profile code which this does.
Also, I added all the logic to the `UpdateProfiles` enum class so we can
keep everything in one place and not have duplicate code and keep
everything into a single for loop.

Original PR: https://github.com/truenas/middleware/pull/16861

---------

Co-authored-by: caleb <yocalebo@gmail.com>