Remove global 2fa check to unset 2FA token
This commit removes the requirement for global 2FA to be disabled
before unsetting a user's 2FA token. The reason to remove this is
that otherwise there is no mechanism for a system administrator
to completely remove 2FA for an existing user without toggling
the global 2FA settings and downgrading system-wide security.

(cherry picked from commit 25874de1ee90def6e0ba258b306075deb399372f)