NAS-137029 / 26.04 / Remove global 2fa check to unset 2FA token (#16893)
This commit removes the requirement for global 2FA to be disabled before
unsetting a user's 2FA token. The reason to remove this is that
otherwise there is no mechanism for a system administrator to completely
remove 2FA for an existing user without toggling the global 2FA settings
and downgrading system-wide security.