NAS-137029 / 25.10-RC.1 / Remove global 2fa check to unset 2FA token (by anodos325) (#16894)
This commit removes the requirement for global 2FA to be disabled before
unsetting a user's 2FA token. The reason to remove this is that
otherwise there is no mechanism for a system administrator to completely
remove 2FA for an existing user without toggling the global 2FA settings
and downgrading system-wide security.

Original PR: https://github.com/truenas/middleware/pull/16893

Co-authored-by: Andrew Walker <awalker@ixsystems.com>