NAS-137113 / 26.04 / Fix start/end in temperature aggregation (#16947)
## Problem

When using the endpoint `disk.temperature_agg`, we were calculating
start/end time in the future with doing roughly the following:
```
start = time right now
end = time right now + days (usually 7)
```

Which meant that we asked data from netdata from right now to 7 days in
the future which is not good. Netdata in this case does not error out
but roughly returns the datapoints of the last 24h in such cases.

## Solution

Make sure we use the correct start/end timestamps to reflect reality and
ask about historical data as requested by the consumer.