NAS-137126 / 25.10-RC.1 / Fix proper disk syncing in Xen VMs (by Qubad786) (#16962)
## Problem

When truenas is virtualized using xen, disks can appear in the system as
`xvdb` etc and in this case their disk identifier shows up as `''` and
eventually we cannot have more then one disk in the disk table because
we have a UNIQUE constraint on identifiers.

Why identifier shows up as `''` because when trying to gauge identifier,
it gets available disks and if the disk in question is not found there,
it skips it (it is not nice but we are going to redo this logic
separately).

## Solution

Make sure the regex we have which covers the disks we support actually
accounts for xen disks as well.

Original PR: https://github.com/truenas/middleware/pull/16960

---------

Co-authored-by: M. Rehan <mrehanlm93@gmail.com>
Co-authored-by: Caleb St. John <30729806+yocalebo@users.noreply.github.com>