Relax validation of email address in query results
This commit switches email address validation so that the more
restrictive checks only occur when creating / updating accounts.
This is to resolve breakage of the users form in the UI for
users who set email addresses under older versions of TrueNAS
where we were not validating email addresses as carefully.