NAS-136901 / 26.04 / Fix `Downloaded update file size mismatch` on unstable connections (#16822)
Server might silently close the connection when downloading the update
file, which will result in an incomplete file being downloaded without
retrying to resume the download. This PR raises an exception if server
closed the connection prematurely, which will cause us to retry the
download a few times before finally giving up.