NAS-136995 / 26.04 / Make directoryservice validation more strict (#17021)
This commit removes ability for users to set a directoryservices.update
request without providing the configuration and credential details if a
service_type has been specified. This ensures that the config details
are always valid. There is already a supported method for users to clear
config if that's what's needed.