NAS-136995 / 25.10-RC.1 / Make directoryservice validation more strict (by anodos325) (#17027)
This commit removes ability for users to set a directoryservices.update
request without providing the configuration and credential details if a
service_type has been specified. This ensures that the config details
are always valid. There is already a supported method for users to clear
config if that's what's needed.

Original PR: https://github.com/truenas/middleware/pull/17021

Co-authored-by: Andrew Walker <awalker@ixsystems.com>