NAS-137171 / 26.04 / Add 'scope' setting to support HA. (#16982)
NFS includes a setting, 'scope'. The purpose of this setting is to
provide a unique and consistent ID that spans a filesystem shared by
multiple controllers, e.g. an HA system.

For some clients, e.g. VMware, to support failover the `scope` setting
must be the same between controllers.

The backports of this depend on updates to `nfs-utils`: 
- https://github.com/truenas/nfs-utils/pull/7

Added CI test which passes on local testing.