NAS-137268 / 26.04 / Add rpc_pipefs setting to satisfy systemd. (#17045)
### DO NOT BACKPORT

For the Halfmoon release the `nfs-utils` repo was updated to trixie. 
The new `nfs-utils` includes changes to systemd and it's management of
NFS and it's parts, e.g. `rpc_pipefs`. One significant change is that
`rpc_pipefs` changed from a `static` file to a `generated` file. The
`generator` is a compiled C module. Inspecting the source exposed a
dependency on the presence of a `pipefs_directory` setting in the
`[general]` section. This PR fixes that and allows the creation of the
rpc_pipefs directory in `/run`. This is the same directory used in the
Debian Trixie release.

Also fixed `test_nonroot_behavior` where using `pytest.raises` would
modify the potentially long exception causing the `assert to fail.
Change to a `try-except` fixed that problem.

### DO NOT BACKPORT