Clear authentication state on logout
There are some situations in which an API client may keep a
websocket connection open after logout and then log back in.
Due to us storing the old authenticated_credentials in the object
associated with the websocket connection, this caused the audit
message to be logged against the last logged-in user account
rather than properly reflecting the authenticated user as
"UNAUTHENTICATED" at that time. This commit clears the
authentication_context (pam handle) and the autheneticated_credentials
on user logout to prevent them from ever being reused.