NAS-137318 / 25.10-RC.1 / Use SSL RNG to generate sensitive IDs (by anodos325) (#17083)
This commit switches middleware to using SSL to generate the random
bytes needed for uuids that are used for session IDs or other
security-adjacent areas. Middleware is also slightly adjusted so that we
generate a unique per-login uuid as well to differentiate between logins
on the same websocket connection.

Original PR: https://github.com/truenas/middleware/pull/17074

Co-authored-by: Andrew Walker <awalker@ixsystems.com>