NAS-137266 / 26.04 / Make sure create paths properly get executed (#17057)
## Problem

`create_paths` in system dataset spec definition is applied EACH time
system dataset setup is executed. The problem happens when a new
sub-dataset is defined in system dataset setup and it has `create_paths`
defined, by default these datasets have legacy mountpoint set. Which
means if the create paths refer to a path on the just created dataset,
that would get created on the _parent_ dataset instead of the actual
dataset because the dataset which was created just right now had not
been mounted.

## Solution

Make sure that we execute create relevant paths endpoint when the
dataset in question has already mounted.
One of the key changes here apart from having it execute once the
dataset is already mounted is that this will not be executed if the
dataset in question is already mounted. Earlier in `setup_datasets`
func, it gets called regardless and it always runs create paths
function.

We could optionally also run this even if the dataset is already mounted
but it would/should be redundant in that case.