NAS-137393 / 26.04 / fix system dataset setup on HA (#17111)
In https://github.com/truenas/middleware/pull/16950, the `pool.query`
call was removed. It was replaced with our newly minted (and more
efficient) zfs.resource.query_impl call. However, there was subtly
implicit behavior with `pool.query`. I would argue, however, that the
implicit behavior with `pool.query` actually was an ORIGINAL bug that
was never caught by previous developers because of how unintuitive it
is.

Let me try to explain. Currently there is a deadlock that is occurring
on the backup controller because this specific code is executed. This
specific code is executed because, on the standby controller, the zpools
DO NOT EXIST (by design). So when reading this logic it makes sense. If
the zpool doesn't exist and the config of the system dataset doesn't
exist on the boot pool then we re-configure the system dataset.

The problem is that `sysdataset.setup` method calls
`systemdataset.update` method which turns around and calls
`sysdataset.setup` This means we have a job, with a lock, that calls a
method that calls the exact same job which causes the deadlock.

The reason why this regressed is painful..... The previous check ran `if
not pool.query(...` but this code was NEVER executing. The reason why it
was never executing is because `pool.query` ALWAYS returned an object
because the underlying zpool that we were querying existed in the
database. So the more accurate understanding of what the the `if not
pool.query` check did is transliterated to something like "if pool does
not exist in our database AND does not exist at OS".

BUT EVEN THEN this code is broken because the object returned from
`pool.query` was not checked to make sure the pool was actually
`ONLINE`.

Sigh....I stared into this abyss and it stared back. Now I want to go
cut yards for a living and leave computers behind.

ANYWAYS, the fix here is to revert back to the original behavior of
`pool.query` by ensuring that the pool we're checking DOES NOT exist in
the database before checking to see if it's imported.